Revert "feat(uring): pass raw descriptors directly and split code in several files"
This reverts commit bb1475d7e1480f5ba13ff0de5138ee6de00e51bc.
