+++ mirror_fs/src/fs/commit_impl.rs
@@ -1,3 +1,5 @@
+use std::os::unix::io::FromRawFd;
+
 use libc;
 use nfs_mamont::vfs::commit;
 
@@ -32,12 +34,15 @@ impl commit::Commit for MirrorFS {
                     });
                 }
             };
+            let file = unsafe { std::fs::File::from_raw_fd(fd) };
             if let Err(error) = uring.fsync(fd, false).await {
+                drop(file);
                 return Err(commit::Fail {
                     error: Self::io_error_to_vfs(&error),
                     file_wcc: self.wcc_data(&path, before).await,
                 });
             }
+            drop(file);
         } else {
             return Err(commit::Fail {
                 error: vfs::Error::IO,
